Fix: server IT tests need AbstractOnServerIT
