+++ src/androidTest/java/com/nextcloud/client/account/UserAccountManagerImplTest.java
@@ -6,7 +6,7 @@ import android.os.Bundle;
 
 import com.nextcloud.client.preferences.AppPreferences;
 import com.nextcloud.client.preferences.AppPreferencesImpl;
-import com.owncloud.android.AbstractIT;
+import com.owncloud.android.AbstractOnServerIT;
 import com.owncloud.android.lib.common.accounts.AccountUtils;
 
 import org.junit.Before;
@@ -17,7 +17,7 @@ import static junit.framework.TestCase.assertNull;
 import static junit.framework.TestCase.assertTrue;
 import static org.junit.Assert.assertFalse;
 
-public class UserAccountManagerImplTest extends AbstractIT {
+public class UserAccountManagerImplTest extends AbstractOnServerIT {
 
     private AccountManager accountManager;
