+++ src/androidTest/java/com/owncloud/android/UploadIT.java
@@ -37,10 +37,8 @@ import com.owncloud.android.utils.FileStorageUtils;
 
 import org.junit.After;
 import org.junit.Test;
-import org.junit.runner.RunWith;
 
 import androidx.annotation.NonNull;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import static junit.framework.TestCase.assertFalse;
 import static junit.framework.TestCase.assertTrue;
@@ -49,8 +47,7 @@ import static junit.framework.TestCase.assertTrue;
  * Tests related to file uploads
  */
 
-@RunWith(AndroidJUnit4.class)
-public class UploadIT extends AbstractIT {
+public class UploadIT extends AbstractOnServerIT {
     private static final String FOLDER = "/testUpload/";
 
     private UploadsStorageManager uploadsStorageManager =
